refactor: rename location
