+++ cognite_toolkit/_cdf_tk/commands/_migrate/migration_io.py
@@ -27,7 +27,7 @@ from cognite_toolkit._cdf_tk.client.resource_classes.three_d import (
 )
 from cognite_toolkit._cdf_tk.commands._migrate.data_classes import ThreeDMigrationRequest
 from cognite_toolkit._cdf_tk.constants import MISSING_EXTERNAL_ID
-from cognite_toolkit._cdf_tk.cruds._resource_cruds.streams import StreamIO
+from cognite_toolkit._cdf_tk.cruds._resource_ios.streams import StreamIO
 from cognite_toolkit._cdf_tk.exceptions import ToolkitNotImplementedError, ToolkitValueError
 from cognite_toolkit._cdf_tk.storageio import (
     AnnotationIO,
