+++ cognite_toolkit/_cdf_tk/cruds/_base_cruds.py
@@ -233,12 +233,12 @@ class ResourceIO(Loader, ABC, Generic[T_Identifier, T_RequestResource, T_Respons
         if parent_ids is not None and not self.parent_resource:
             return []
         if space is not None:
-            from ._resource_cruds.datamodel import SpaceCRUD
+            from ._resource_ios.datamodel import SpaceCRUD
 
             if SpaceCRUD not in self.dependencies:
                 return []
         if data_set_external_id is not None:
-            from ._resource_cruds.data_organization import DataSetsIO
+            from ._resource_ios.data_organization import DataSetsIO
 
             if DataSetsIO not in self.dependencies:
                 return []
