+++ cognite_toolkit/_cdf_tk/cruds/_resource_ios/industrial_tool.py
