+++ cognite_toolkit/_cdf_tk/cruds/_resource_ios/signal_subscription.py
@@ -15,12 +15,12 @@ from cognite_toolkit._cdf_tk.client.resource_classes.signal_subscription import 
     SignalSubscriptionResponse,
 )
 from cognite_toolkit._cdf_tk.cruds._base_cruds import ResourceIO
-from cognite_toolkit._cdf_tk.cruds._resource_cruds.hosted_extractors import (
+from cognite_toolkit._cdf_tk.cruds._resource_ios.hosted_extractors import (
     HostedExtractorDestinationIO,
     HostedExtractorSourceIO,
 )
-from cognite_toolkit._cdf_tk.cruds._resource_cruds.signal_sink import SignalSinkIO
-from cognite_toolkit._cdf_tk.cruds._resource_cruds.workflow import WorkflowIO
+from cognite_toolkit._cdf_tk.cruds._resource_ios.signal_sink import SignalSinkIO
+from cognite_toolkit._cdf_tk.cruds._resource_ios.workflow import WorkflowIO
 from cognite_toolkit._cdf_tk.yaml_classes import SignalSubscriptionYAML
 from cognite_toolkit._cdf_tk.yaml_classes.signal_subscription import (
     HostedExtractorsFilterYAML,
