+++ cognite_toolkit/_cdf_tk/cruds/_resource_ios/three_d_model.py
