+++ cognite_toolkit/_cdf_tk/data_classes/_build_variables.py
@@ -8,7 +8,7 @@ from functools import cached_property
 from pathlib import Path
 from typing import Any, Literal, SupportsIndex, overload
 
-from cognite_toolkit._cdf_tk.cruds._resource_cruds.transformation import TransformationIO
+from cognite_toolkit._cdf_tk.cruds._resource_ios.transformation import TransformationIO
 from cognite_toolkit._cdf_tk.data_classes._module_directories import ModuleLocation
 from cognite_toolkit._cdf_tk.exceptions import ToolkitValueError
 from cognite_toolkit._cdf_tk.feature_flags import Flags
