+++ tests/test_integration/test_cruds/test_resource_crud_containers.py
@@ -17,7 +17,7 @@ from cognite_toolkit._cdf_tk.client.resource_classes.three_d import (
 )
 from cognite_toolkit._cdf_tk.client.resource_classes.timeseries import TimeSeriesRequest
 from cognite_toolkit._cdf_tk.cruds import ContainerCRUD, RawTableCRUD, TimeSeriesCRUD
-from cognite_toolkit._cdf_tk.cruds._resource_cruds.three_d_model import ThreeDModelCRUD
+from cognite_toolkit._cdf_tk.cruds._resource_ios.three_d_model import ThreeDModelCRUD
 from tests.test_integration.constants import RUN_UNIQUE_ID
