+++ tests/test_unit/test_cdf_tk/test_commands/test_buildv2/test_command.py
@@ -22,8 +22,8 @@ from cognite_toolkit._cdf_tk.commands.build_v2.data_classes._types import Absolu
 from cognite_toolkit._cdf_tk.constants import MODULES
 from cognite_toolkit._cdf_tk.cruds import FileMetadataCRUD, SearchConfigIO, SpaceCRUD
 from cognite_toolkit._cdf_tk.cruds._base_cruds import ResourceIO
-from cognite_toolkit._cdf_tk.cruds._resource_cruds.datamodel import DataModelIO, ViewIO
-from cognite_toolkit._cdf_tk.cruds._resource_cruds.workflow import WorkflowIO
+from cognite_toolkit._cdf_tk.cruds._resource_ios.datamodel import DataModelIO, ViewIO
+from cognite_toolkit._cdf_tk.cruds._resource_ios.workflow import WorkflowIO
 from cognite_toolkit._cdf_tk.exceptions import ToolkitError, ToolkitValueError
 from cognite_toolkit._cdf_tk.rules._dependencies import DependencyRuleSet
